# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: TPX LE proxy
  description: >
    This is an interface proxy between ThingPark X Location Engine and different types of LoRaWAN Network Servers.<br />
    An example application server utilising this API is available <a href='https://nano-things.net/tpxle-proxy/app' target='_blank'>here</a>.<br />
    The source code of this API is published on Github <a href='https://github.com/norbertherbert/tpxle-proxy' target='_blank'>here</a>.
servers:
  - url: 'https://nano-things.net/tpxle-proxy/latest/api'
    description: Demo server
  - url: 'http://localhost:8086/tpxle-proxy/latest/api'
    description: Local test server
...
paths:
  /uplink_location_solver:
    post:
      tags:
        - From Location Solver
      summary: Receive resolved location from ThingPark Location Solver
      description: This endpoint receives resolved location from ThingPark Location Solver and saves it in the servers DB.
      x-swagger-router-controller: fromLocationSolver
      operationId: saveResolvedLocation
      requestBody:
        description: The message body sent by the ThingPark Location Solver.
        x-name: body
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LsResolvedLocation"
      responses:
        '200':
          description: \[200 OK\] The message was successfully received and forwarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'  
  /resolved_received:
    get:
      tags:
        - From Location Solver
      summary: List resolved locations received from Location Solver.
      description: This endpoint lists resolved locations recently received from ThingPark Location Solver.
      x-swagger-router-controller: fromLocationSolver
      operationId: listResolvedLocations
      parameters:
        - name: deviceEUI
          description: Use this parameter to filter out one device with a specific DevEUI.
          in: query
          schema:
            type: string
      responses:
        '200':
          description: \[200 OK\] The list has been retreived
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LsResolvedLocation"
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'  
  /uplink_actility:
    post:
      tags:
        - From Network Servers
      summary: Receive uplinlk message from Actility Network Server
      description: This endpoint receives an uplink message from Actility Network Server, transforms it to the right format and forwards it to ThingPark Location Solver.
      x-swagger-router-controller: fromNetworkServer
      operationId: uplinkActility
      parameters:
        - name: AS_ID
          in: query
          description: Application Server ID (only reported when the authentication is activated). AS_ID is not reported when the Uplink/downlink authentication is not activated in the AS profile.
          schema:
            type: string
        - name: LrnDevEui
          in: query
          description: Device DevEUI
          schema:
            type: string
        - name: LrnFPort
          in: query
          description: LoRaWAN™ port number
          schema:
            type: number
        - name: LrnInfos
          in: query
          description: Service profile name used to route the packet
          schema:
            type: string
        - name: Time
          in: query
          description: ISO 8601 timestamp associated to the generation of the HTTP request by the LRC (only reported when the authentication is activated). Time is not reported when the Uplink/downlink authentication is not activated in the AS profile.
          schema:
            type: string
            format: date-time
        - name: Token
          in: query
          description: Security token generated by the LRC (only reported when the authentication is activated). Token is not reported when the Uplink/downlink authentication is not activated in the AS profile.
          schema:
            type: string
      requestBody:
        description: The message body sent by the Network Server.
        x-name: body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NsActility'
          application/xml:
            schema:
              $ref: '#/components/schemas/NsActility'
      responses:
        '200':
          description: \[200 OK\] The message was successfully received.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /uplink_ttn:
    post:
      tags:
        - From Network Servers
      summary: Receive uplinlk message from TTN Network Server
      description: This endpoint receives an uplink message from TTN Network Server, transforms it to the right format and forwards it to ThingPark Location Solver.
      x-swagger-router-controller: fromNetworkServer
      operationId: uplinkTTN
      requestBody:
        description: The message body sent by the Network Server.
        x-name: body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NsTTN'
      responses:
        '200':
          description: \[200 OK\] The message was successfully received.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /uplink_loriot:
    post:
      tags:
        - From Network Servers
      summary: Receive uplinlk message from Loriot Network Server
      description: This endpoint receives an uplink message from Loriot Network Server, transforms it to the right format and forwards it to ThingPark Location Solver.
      x-swagger-router-controller: fromNetworkServer
      operationId: uplinkLoriot
      requestBody:
        description: The message body sent by the Network Server.
        x-name: body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NsLoriot'
      responses:
        '200':
          description: \[200 OK\] The message was successfully received.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /uplink_kerlink:
    post:
      tags:
        - From Network Servers
      summary: Receive uplinlk message from Kerlink Network Server
      description: This endpoint receives an uplink message from Kerlink Network Server, transforms it to the right format and forwards it to ThingPark Location Solver.
      x-swagger-router-controller: fromNetworkServer
      operationId: uplinkKerlink
      requestBody:
        description: The message body sent by the Network Server.
        x-name: body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NsKerlink'
      responses:
        '200':
          description: \[200 OK\] The message was successfully received.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /received_decoded:
    get:
      tags:
        - From Network Servers
      summary: List messages recently received from Network Servers.
      description: This endpoint lists locally decoded messages recently received from Network Servers.
      x-swagger-router-controller: fromNetworkServer
      operationId: listDecodedMessages
      parameters:
        - name: devEUI
          description: Use this parameter to filter out one device with a specific DevEUI.
          in: query
          schema:
            type: string
      responses:
        '200':
          description: \[200 OK\] The list has been retreived
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AsDecodedMessage"
        '400':
          description: \[400 Bad Request\] Invalid parameters supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: \[403 Forbidden\] Authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: \[500 Internal Server Error\] The server encountered an unexpected
            condition which prevented it from fulfilling the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'  
components:
  schemas:
    LsResolvedLocation:
      type: object
      properties:
        dxProfileId:
          type: string 
          example: poc-api
        customerId: 
          type: string 
          example: '100002659'
        deviceEUI: 
          type: string 
          example: 20635F00C50000AF
        time:
          type: string
          format: date-time
          example: '2019-04-03T08:11:15.335000+00:00'
        coordinates:
          type: array
          items: 
            type: number
          example: [19.047703, 47.46676, 146]
        age: 
          type: number 
          example: 0
        validityState: 
          type: string 
          example: NEW
        horizontalAccuracy: 
          type: number 
          example: 65
        processedFeed:
          type: object
          properties:
            payloadEncoded: 
              type: string 
              example: 032c946e19004432c832f9bfbe48f8b37c935fbd38437d8aafe4b93a431d8aafe4b8
            sequenceNumber: 
              type: number 
              example: 9233
            dynamicMotionState: 
              type: string 
              example: MOVING
            temperatureMeasure: 
              type: number 
              example: 11.6470585
            processedPacket:
              type: object
              properties:
                RSSI: 
                  type: number
                  example: -89
                baseStationId: 
                  type: string
                  example: 004A0DAA
                antennaCoordinates:
                  type: array
                  items:
                    type: number
                  example: [0, 0]
        rawPosition:
          type: object
          properties:
            rawPositionType: 
              type: string 
              example: RawPositionByWifiSolver
            coordinates:
              type: array
              items:
                type: number
              example: [19.047703, 47.46676]
            bssidCount: 
              type: number 
              example: 4
        resolvedTracker:
          type: object
          properties:
            firmwareVersion:
              type: string 
              example: 1.7.1
            messageType: 
              type: string 
              example: Position Message
            trackingMode: 
              type: string 
              example: MOTION_TRACKING
            gpsScanMode: 
              type: string 
              example: UNKNOWN
            sensorMode: 
              type: string 
              example: UNKNOWN
            batteryLevel: 
              type: number 
              example: 99
            activityCount: 
              type: number 
              example: -1
...
    NsTTN:
      type: object
      properties:
        app_id:
          description: Same as in the topic
          type: string
          example: 'my-app-id'
        dev_id:
          description: Same as in the topic
          type: string
          example: 'my-dev-id'
        hardware_serial:
          description: 'In case of LoRaWAN: the DevEUI'
          type: string
          example: '0102030405060708'
        port:
          description: LoRaWAN FPort
          type: integer
          example: 1
        counter:
          description: LoRaWAN frame counter
          type: integer
          example: 2
        is_retry:
          description: Is set to true if this message is a retry (you could also detect this from the counter)
          type: boolean
          example: false
        confirmed:
          description: Is set to true if this message was a confirmed message
          type: boolean
          example: false
        payload_raw:
          description: 'Base64 encoded payload: [0x01, 0x02, 0x03, 0x04]'
          type: string
          example: 'AQIDBA=='
        payload_fields:
          description: Object containing the results from the payload functions - left out when empty
          type: object
          example: {}

        metadata:
          description: ''
          type: object
          properties:
            time:
              description: Time when the server received the message
              type: string
              format: date-time
              example: '1970-01-01T00:00:00Z'
            frequency:
              description: Frequency at which the message was sent
              type: number
              example: 868.1
            modulation:
              description: Modulation that was used - LORA or FSK
              type: string
              example: 'LORA'
            data_rate:
              description: Data rate that was used - if LORA modulation
              type: string
              example: 'SF7BW125'
            bit_rate:
              description: Bit rate that was used - if FSK modulation
              type: number
              example: 50000
            coding_rate:
              description: Coding rate that was used
              type: string
              example: '4/5'
            gateways:
              type: array
              items:
                type: object
                properties:
                  gtw_id: 
                    description: EUI of the gateway
                    type: string
                    example: 'ttn-herengracht-ams'
                  timestamp: 
                    description: Timestamp when the gateway received the message
                    type: number
                    example: 12345
                  time: 
                    description: Time when the gateway received the message - left out when gateway does not have synchronized time
                    type: string
                    format: date-time
                    example: '1970-01-01T00:00:00Z'
                  channel: 
                    description: Channel where the gateway received the message
                    type: number
                    example: 0
                  rssi: 
                    description: Signal strength of the received message
                    type: number
                    example: -25
                  snr: 
                    description: Signal to noise ratio of the received message
                    type: number
                    example: 5
                  rf_chain: 
                    description: RF chain where the gateway received the message
                    type: integer
                    example: 0
                  latitude: 
                    description: Latitude of the gateway reported in its status updates
                    type: number
                    example: 52.1234
                  longitude: 
                    description: Longitude of the gateway
                    type: number
                    example: 6.1234
                  altitude: 
                    description: Altitude of the gateway
                    type: number
                    example: 6
            latitude:
              description: Latitude of the device
              type: number
              example: 52.2345
            longitude:
              description: Longitude of the device
              type: number
              example: 6.2345
            altitude:
              description: Altitude of the device
              type: number
              example: 2
        downlink_url:
          description: ''
          type: string
          example: 'https://integrations.thethingsnetwork.org/ttn-eu/api/v2/down/my-app-id/my-process-id?key=ttn-account-v2.secret'
    NsLoriot:
      type: object
      properties:
        cmd: 
          description: identifies type of message, rx = uplink message
          type: string
          example: 'rx'
        EUI: 
          description: device EUI, 16 hex digits (without dashes)
          type: string
          example: '0011223344556677'
        ts: 
          description: server timestamp as number (miliseconds from Linux epoch)
          type: integer
          example: 1555847833236
        ack: 
          description: NEW! acknowledgement flag as set by device
          type: boolean
          example: false
        fcnt: 
          description: frame counter, a 32-bit number
          type: integer
          example: 118
        port: 
          description: port as sent by the end device
          type: integer
          example: 17
        encdata: 
          description: data payload (APPSKEY encrypted hex string) only present if APPSKEY is assigned to device
          type: string
          example: 'af234b'
        data: 
          description: data payload (decrypted, plaintext hex string) only present if APPSKEY is not assigned to device
          type: string
          example: 'aabbcc11'
        freq: 
          description: radio frequency at which the frame was received, in Hz
          type: integer
          example: 868100000
        dr: 
          description: radio data rate - spreading factor, bandwidth and coding rate e.g. SF12 BW125 4/5
          type: string
          example: 'SF12 BW125 4/5'
        rssi: 
          description: frame rssi, in dBm, as integer number
          type: integer
          example: -97
        snr: 
          description: frame snr, in dB, one decimal place
          type: number
          example: -7.25
    NsKerlink:
      type: object
      properties:
        id:
          type: string
          example: '5da03ee3a96ab200018d5f3c'
        endDevice:
          type: object
          properties:
            devEui:
              type: string
              example: '20635F0108001158'
            devAddr:
              type: string
              example: '0204E05E'
            cluster:
              type: object
              properties:
                id:
                  type: integer
                  example: 435
        fPort:
          type: integer
          example: 17
        fCntDown:
          type: integer
          example: 1
        fCntUp:
          type: integer
          example: 9
        adr:
          type: boolean
          example: true
        confirmed:
          type: boolean
          example: false
        encrypted:
          type: boolean
          example: false
        payload:
          type: string
          example: '0a50009e0000'
        recvTime:
          type: integer
          example: 1570782946959
        classB:
          type: boolean
          example: false
        delayed:
          type: boolean
          example: false
        ulFrequency:
          type: number
          example: 868.5
        modulation:
          type: string
          example: 'LORA'
        dataRate:
          type: string
          example: 'SF10BW125'
        codingRate:
          type: string
          example: '4/5'
        gwCnt:
          type: integer
          example: 1
        gwInfo:
          type: array
          items:
            type: object
            properties:
              gwEui:
                type: string 
                example: '7276FF0039030872'
              rfRegion: 
                type: string 
                example: 'EU868'
              rssi: 
                type: integer 
                example: -47
              snr: 
                type: integer 
                example: 7.5
              channel: 
                type: integer 
                example: 7
              radioId: 
                type: integer 
                example: 1
    AsDecodedMessage:
      type: object
      properties:
        devEUI:
          type: string
        fPort:
          type: integer
        time:
          type: string
          format: date-time
        payloadHex:
          type: string
        payloadDecoded:
          type: object
    MessageResponse:
      type: object
      properties:
        message:
          type: object
          properties:
            text:
              type: string
            code:
              type: integer